refactor(components/molecule/autosuggest): api tab
